requirements.txt
flask==3.0.0
pyserial==3.5
google-generativeai==0.3.2
python-dateutil==2.8.2
numpy==1.24.3
pandas==2.1.4
scikit-learn==1.3.2
python-dotenv==1.0.0
requests==2.31.0
Werkzeug==3.0.1
gunicorn==21.2.0
Flask-Cors==4.0.0 